board/axxia: Use the Correct Pointer Conversion for NCA

+++ include/configs/axm5600.h
@@ -1538,7 +1538,7 @@ extern volatile unsigned long long *scratch;
 void axxia_display_clocks(void);
 int clocks_init(void);
 int voltage_init(void);
-int pciesrio_init(unsigned long);
+int pciesrio_init(unsigned int);
 int is_sbb_enabled(int);
 int sbb_verify_image(void *, void *, int, int, int);
 extern unsigned int pfuse;
